+++ test/libsolidity/smtCheckerTests/functions/functions_identity_as_tuple.sol
@@ -12,3 +12,4 @@ contract C
 }
 
 // ----
+// Warning: (153-156): Assertion checker does not yet implement tuples and inline arrays.