+++ Changelog.md
@@ -6,6 +6,7 @@ Language Features:
 Compiler Features:
  * Inline Assembly: Improve error messages around invalid function argument count.
  * Code Generator: Use binary search for dispatch function if more efficient. The size/speed tradeoff can be tuned using ``--optimize-runs``.
+ * SMTChecker: Support mathematical and cryptographic functions in an uninterpreted way.
  * Type Checker: Add an additional reason to be displayed when type conversion fails.